+++ packages/cli/src/cli.ts
@@ -5,6 +5,8 @@ import { repo as repoCommand, install as installCommand } from './repo/command';
 import executeCommand from './execute/command';
 import compileCommand from './compile/command';
 import testCommand from './test/command';
+import docgenCommand from './docgen/command';
+import docsCommand from './docs/command';
 import { Opts } from './commands';
 
 export const cmd = yargs(hideBin(process.argv))
@@ -13,11 +15,18 @@ export const cmd = yargs(hideBin(process.argv))
   .command(installCommand) // allow install to run from the top as well as repo
   .command(repoCommand)
   .command(testCommand)
+  .command(docsCommand)
+  .command(docgenCommand)
   // Common options
   .option('log', {
     alias: ['l'],
     description: 'Set the default log level to none, default, info or debug',
     array: true,
   })
+  .example('openfn execute help', 'Show documentation for the execute command')
+  .example(
+    'openfn docs @openfn/language-common each',
+    'Get more help on the common.each command'
+  )
   .alias('v', 'version')
   .help() as yargs.Argv<Opts>;