+++ packages/cli/test/__monorepo__/packages/common/index.js
@@ -1,0 +1,1 @@
+export const alterState = (f) => (state) => f(state);