+++ packages/runtime-manager/CHANGELOG.md
@@ -1,5 +1,12 @@
 # runtime-manager
 
+## 0.0.18
+
+### Patch Changes
+
+- @openfn/compiler@0.0.19
+- @openfn/runtime@0.0.13
+
 ## 0.0.17
 
 ### Patch Changes